lockout user

+++ Controllers/AdminController.cs
@@ -18,14 +18,33 @@ public class AdminController : Controller
         _userManager = userManager;
     }
 
+    private bool IsUserAuthorized(int departmentId)
+    {
+        var user = _context.Users.FirstOrDefault(u => u.UserName.Equals(User.Identity.Name));
+        return user?.DepartmentId == departmentId;
+    }
+
+    private IActionResult RedirectToHome()
+    {
+        return RedirectToAction("Index", "Home");
+    }
+
     public IActionResult Index()
     {
+        if (!IsUserAuthorized(5))
+        {
+            return RedirectToHome();
+        }
         var users = _context.Users.Include(u => u.Department).ToList();
         return View(users);
     }
 
     public IActionResult Details(string id)
     {
+        if (!IsUserAuthorized(5))
+        {
+            return RedirectToHome();
+        }
         if (id == null)
         {
             return NotFound();
@@ -45,6 +64,10 @@ public class AdminController : Controller
 
     public IActionResult Edit(string id)
     {
+        if (!IsUserAuthorized(5))
+        {
+            return RedirectToHome();
+        }
         if (id == null)
         {
             return NotFound();
@@ -66,6 +89,10 @@ public class AdminController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Edit(string id, User user)
     {
+        if (!IsUserAuthorized(5))
+        {
+            return RedirectToHome();
+        }
         if (id != user.Id)
         {
             return NotFound();
@@ -98,6 +125,10 @@ public class AdminController : Controller
 
     public IActionResult Delete(string id)
     {
+        if (!IsUserAuthorized(5))
+        {
+            return RedirectToHome();
+        }
         var user = _context.Users.Find(id);
         _context.Users.Remove(user);
         _context.SaveChanges();
